+++ src/interfaces/custom-file-template.interface.ts
@@ -1,5 +1,5 @@
 export interface ICustomFileTemplate {
     itemName: string;
     template?: string;
-    templateUrl?: string;
+    templateFile?: string;
 }