+++ changelog.md
@@ -1,14 +1,21 @@
 # NRG Changelog
 
 ## 0.3.0 - Aelinor
-* Solved init bug
-* Added badges on readme
-* Updated package repo data
+
+#### New features:
+* You can now specify a custom template for default items and custom items 
 * Updated json schema (added customFileTemplates)
 * Updated json schema (added templateUrl to customFileTemplates)
 * Updated json schema (added description for fields and examples)
 * Updated json schema TS types
 * Updated json schema TS types (added $schema entry)
+
+#### Bugfixes
+* Solved init bug
+
+#### Minor changes
+* Added badges on readme
+* Updated package repo data
 
 ## 0.2.4 - Aelene
 * Throwing errors for: MissingItemTypeFlagException and unhandled errors.