+++ changelog.md
@@ -2,7 +2,8 @@
 ---
 ### 0.3.0 - Aelinor
 #### New features:
-* You can now specify a custom template for default items and custom items 
+* You can now define a default folders structure to scaffold
+* You can now specify a custom templates for default items and custom items 
 * Updated json schema (added customFileTemplates)
 * Updated json schema (added templateUrl to customFileTemplates)
 * Updated json schema (added description for fields and examples)